support version-based key-value service.

+++ service/tools/utxo/wallet_tool/cpp/utxo_client_tools.cpp
@@ -46,9 +46,10 @@ void ShowUsage() {
 }
 
 void Transfer(UTXOClient* client, int64_t transaction_id,
-              const std::string& address, 
+              const std::string& address,
               const std::vector<std::string>& to_address,
-              const std::vector<int64_t>& values, const std::string& private_key,
+              const std::vector<int64_t>& values,
+              const std::string& private_key,
               const std::vector<std::string>& to_pub_key) {
   if (private_key.empty() || to_pub_key.empty()) {
     printf("no private key or public key\n");
@@ -61,15 +62,16 @@ void Transfer(UTXOClient* client, int64_t transaction_id,
   in->set_out_idx(0);
   nonce += transaction_id;
 
-  for(int i = 0; i < to_address.size();++i){
+  for (size_t i = 0; i < to_address.size(); ++i) {
     UTXOOut* out = utxo.add_out();
     out->set_address(to_address[i]);
     out->set_value(values[i]);
     out->set_pub_key(to_pub_key[i]);
     utxo.set_address(address);
-    utxo.set_sig(resdb::utils::ECDSASignString(private_key,
-                                               address + std::to_string(nonce)));
-    LOG(ERROR)<<"transfer from:"<<address<<" to:"<<to_address[i]<<" value:"<<values[i];
+    utxo.set_sig(resdb::utils::ECDSASignString(
+        private_key, address + std::to_string(nonce)));
+    LOG(ERROR) << "transfer from:" << address << " to:" << to_address[i]
+               << " value:" << values[i];
   }
 
   auto output = client->Transfer(utxo);
@@ -91,34 +93,33 @@ void GetWallet(UTXOClient* client, const std::string& address) {
   LOG(ERROR) << "address:" << address << " get wallet value:" << ret;
 }
 
-std::vector<std::string> ParseString(std::string str){
+std::vector<std::string> ParseString(std::string str) {
   std::vector<std::string> ret;
-  while(true){
+  while (true) {
     size_t pos = str.find(",");
-    if(pos == std::string::npos){
+    if (pos == std::string::npos) {
       ret.push_back(str);
       break;
     }
-    ret.push_back(str.substr(0,pos));
-    str = str.substr(pos+1);
+    ret.push_back(str.substr(0, pos));
+    str = str.substr(pos + 1);
   }
   return ret;
 }
 
-std::vector<int64_t> ParseValue(std::string str){
+std::vector<int64_t> ParseValue(std::string str) {
   std::vector<int64_t> ret;
-  while(true){
+  while (true) {
     size_t pos = str.find(",");
-    if(pos == std::string::npos){
+    if (pos == std::string::npos) {
       ret.push_back(strtoull(str.c_str(), NULL, 10));
       break;
     }
-    ret.push_back(strtoull(str.substr(0,pos).c_str(), NULL, 10));
-    str = str.substr(pos+1);
+    ret.push_back(strtoull(str.substr(0, pos).c_str(), NULL, 10));
+    str = str.substr(pos + 1);
   }
   return ret;
 }
-
 
 int main(int argc, char** argv) {
   if (argc < 3) {
@@ -177,8 +178,8 @@ int main(int argc, char** argv) {
   config.SetClientTimeoutMs(100000);
   UTXOClient client(config);
   if (cmd == "transfer") {
-    Transfer(&client, transaction_id, address, ParseString(to_address), ParseValue(value), private_key,
-             ParseString(to_pub_key));
+    Transfer(&client, transaction_id, address, ParseString(to_address),
+             ParseValue(value), private_key, ParseString(to_pub_key));
   } else if (cmd == "list") {
     GetList(&client, end_id, num);
   } else if (cmd == "wallet") {
